+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Version history
 
 ## 0.1.1
+- Move `Sorbet` status bar item to be part of the `Ruby` language item for more standard behavior. 
 - Removed unnecessary files fom .vsix.
 
 ## 0.1.0